Move Jobs table action bar to its own component, and remove most manual height logic
