+++ src/components/jobsTable/JobsTableActionBar.module.css
@@ -1,0 +1,29 @@
+.actionBar {
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    margin-bottom: 0.5em;
+}
+
+.actionGroup {
+    display: flex;
+    flex-direction: row;
+    column-gap: 0.5em;
+}
+.actionBar .actionGroup:last-child {
+    margin-left: auto;
+}
+
+.actions {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
+
+.actionItem {
+    margin-left: 0.5em;
+    margin-right: 0.5em;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
