+++ src/pages/JobsPage.module.css
@@ -5,21 +5,6 @@
     align-items: center;
 }
 
-.actions {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-}
-
-.actionItem {
-    margin-left: 0.5em;
-    margin-right: 0.5em;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
 .jobsTable {
-    margin-left: 0.5em;
-    margin-right: 0.5em;
+    margin: 0.5em;
 }
